enabling automatic DB cleanup again

+++ src/db.c
@@ -485,23 +485,25 @@ open:
 		debug_end_measurement (DEBUG_DB, "table setup");
 
 		/* Cleanup of DB */
-	
-		/*
-		debug_start_measurement (DEBUG_DB);
-		db_exec ("DELETE FROM items WHERE ROWID NOT IN "
-			 "(SELECT item_id FROM itemsets);");
-		debug_end_measurement (DEBUG_DB, "cleanup lost items");
 
-		debug_start_measurement (DEBUG_DB);
-		db_exec ("DELETE FROM itemsets WHERE item_id NOT IN "
-			 "(SELECT ROWID FROM items);");
-		debug_end_measurement (DEBUG_DB, "cleanup lost itemset entries");
+		if (initial) {	
+			debug0 (DEBUG_DB, "Checking for items not referenced in table 'itemsets'...\n");
+			db_exec ("DELETE FROM items WHERE ROWID NOT IN "
+				 "(SELECT item_id FROM itemsets);");
+				 
+			debug0 (DEBUG_DB, "Checking for invalid item ids in table 'itemsets'...\n");
+			db_exec ("CREATE TEMP TABLE tmp_id ( id );");
+			db_exec ("INSERT INTO tmp_id SELECT item_id FROM itemsets WHERE item_id NOT IN (SELECT ROWID FROM items);");
+			/* limit to 1000 items as it is very slow */
+			db_exec ("DELETE FROM itemsets WHERE item_id IN (SELECT id FROM tmp_id LIMIT 1000);");
+			db_exec ("DROP TEMP TABLE tmp_id;");
 
-		debug_start_measurement (DEBUG_DB);
-		db_exec ("DELETE FROM itemsets WHERE comment = 0 AND node_id NOT IN "
-	        	 "(SELECT node_id FROM subscription);");
-		debug_end_measurement (DEBUG_DB, "cleanup lost node entries");
-		*/
+			debug0 (DEBUG_DB, "Checking for items without a subscription...\n");
+			db_exec ("DELETE FROM itemsets WHERE comment = 0 AND node_id NOT IN "
+		        	 "(SELECT node_id FROM subscription);");
+				 
+			debug0 (DEBUG_DB, "DB cleanup finished. Continuing startup.\n");
+		}
 	}
 	
 	/* prepare statements */
